+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [NEXT]
 -  Add SpriteAnimation extension: method `asWidget`.(You can use this to SpriteAnimation or Future<SpriteAnimation>)
 -  Add Sprite extension: method `asWidget`.(You can use this to Sprite or Future<Sprite>)
+-  Add Suport to ImageLayer in map built by Tiled. [issue 76](https://github.com/RafaelBarbosatec/bonfire/issues/76)
 
 # 1.11.1
 - Fix problem render Map.